refactor: 날짜 이동하는 함수 커스텀 훅으로 분리
