+++ src/containers/group/groupId/tasklist/Tasks.tsx
@@ -5,11 +5,11 @@ import useQueryParameter from '@/hooks/useQueryParameter';
 import useTaskLists from '@/hooks/useTaskLists';
 import useTasks from '@/hooks/useTasks';
 import getMonthDay from '@/utils/getMonthDay';
-import { useQueryClient } from '@tanstack/react-query';
 import clsx from 'clsx';
 import Link from 'next/link';
 import { useState } from 'react';
 import TaskAddButton from '../TaskAddButton';
+import useDayNavigation from '../useDayNavigation';
 import Task from './Task';
 import TodoAddButton from './TodoAddButton';
 
@@ -23,30 +23,14 @@ function Tasks() {
 
   const { taskLists, isLoading: isListLoading } = useTaskLists(groupId);
 
-  const queryClient = useQueryClient();
+  const handleNavigateDay = useDayNavigation(setPickDate);
 
   const handlePreviousDay = () => {
-    setPickDate((prevDate) => {
-      const newDate = new Date(prevDate); // 이전 상태 복사
-      newDate.setDate(newDate.getDate() - 1); // 하루 전날로 변경
-      queryClient.invalidateQueries({
-        queryKey: ['getTasks', groupId, taskListId, getMonthDay(newDate)],
-      });
-
-      return newDate;
-    });
+    handleNavigateDay('prev');
   };
 
   const handleNextDay = () => {
-    setPickDate((prevDate) => {
-      const newDate = new Date(prevDate); // 이전 상태 복사
-      newDate.setDate(newDate.getDate() + 1); // 하루 전날로 변경
-      queryClient.invalidateQueries({
-        queryKey: ['getTasks', groupId, taskListId, getMonthDay(newDate)],
-      });
-
-      return newDate;
-    });
+    handleNavigateDay('next');
   };
 
   if (isLoading) return <>임시 로딩중 ... </>;
